requirements.txt
indic-nlp-library==0.81
streamlit==0.68.0